+++ src/pages/FilesUpload/Direct/index.js
@@ -1,4 +1,5 @@
 import axios from 'axios';
+import Cookies from 'js-cookie';
 import { useState } from 'react';
 
 import { BASE_BACKEND_URL } from 'config/urls';
@@ -20,24 +21,72 @@ const DirectUploadExample = () => {
     return config;
   };
 
-  const uploadFile = ({ file }) => {
+  const getCSRFConfig = () => {
+    const csrf = Cookies.get('csrftoken');
+
+    const config = getConfig();
+
+    return {
+      ...config,
+      headers: {
+        ...config.headers,
+        'X-CSRFToken': csrf
+      }
+    };
+  };
+
+  const fileGeneratePresignedPost = ({ fileName, fileType }) => {
+    return axios.post(
+      `${BASE_BACKEND_URL}/api/files/upload/direct/start/`,
+      { file_name: fileName, file_type: fileType },
+      getCSRFConfig()
+    );
+  };
+
+  const uploadFile = ({ data, file }) => {
     const postData = new FormData();
+
+    for (const key in data?.fields) {
+      postData.append(key, data.fields[key]);
+    }
 
     postData.append('file', file);
 
-    const url = `${BASE_BACKEND_URL}/api/files/upload/direct/`;
-    return axios.post(url, postData, getConfig());
+    let postParams = getCSRFConfig();
+
+    // If we're uploading to S3, detach the authorization cookie.
+    if (data?.fields) {
+      postParams = {};
+    }
+
+    return axios
+      .post(data.url, postData, postParams)
+      .then(() => Promise.resolve({ fileId: data.id }));
+  };
+
+  const verifyUpload = ({ data }) => {
+    return axios.post(
+      `${BASE_BACKEND_URL}/api/files/upload/direct/finish/`,
+      { file_id: data.id },
+      getCSRFConfig()
+    );
   };
 
   const onInputChange = (event) => {
     const file = event.target.files[0];
 
     if (file) {
-      uploadFile({ file })
-        .then((response) => {
-          console.log(response);
-          setMessage('File upload completed!');
-        })
+      fileGeneratePresignedPost({
+        fileName: file.name,
+        fileType: file.type
+      })
+        .then((response) =>
+          uploadFile({ data: response.data, file })
+            .then(() => verifyUpload({ data: response.data }))
+            .then(() => {
+              setMessage('File upload completed!');
+            })
+        )
         .catch((error) => {
           setMessage('File upload failed!');
         });
@@ -46,7 +95,7 @@ const DirectUploadExample = () => {
 
   return (
     <div>
-      <h1>Direct upload example!</h1>
+      <h1>Direct upload</h1>
       <div>Select files to upload:</div>
 
       <input id="input" type="file" onChange={onInputChange} />
