+++ src/pages/FilesUpload/index.js
@@ -1,13 +1,12 @@
-// See FilesUploadJsExample file if you want to have this code in pure html
-
+import StandardUploadExample from './Standard';
 import DirectUploadExample from './Direct';
-import FileUploadReactExample from './FileUploadReactExample';
 
 const UploadFiles = () => {
   return (
     <div>
-      {false && <DirectUploadExample />}
-      {true && <FileUploadReactExample />}
+      <StandardUploadExample />
+      <hr />
+      <DirectUploadExample />
     </div>
   );
 };
